# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "中台api",
        "description": "这是演示服务，该文档提供了演示swagger api的功能",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "开发环境"
        },
        {
            "url": "http://uat",
            "description": "uat环境"
        }
    ],
    "paths": {
...
        "/order/v1/detail/{orderNo}": {
            "get": {
                "tags": [
                    "订单中心"
                ],
                "summary": "获取订单详情",
                "description": "获取订单详情",
                "operationId": "orderDetail",
                "parameters": [
                    {
                        "name": "orderNo",
                        "in": "path",
                        "description": "主订单编号",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功示例",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "请求是否成功，结果为true表示成功，false表示失败",
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "description": "状态码。0为成功",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "消息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "orderNo": {
                                                    "description": "订单编号",
                                                    "type": "string"
                                                },
                                                "bankNo": {
                                                    "description": "银行订单号",
                                                    "type": "string"
                                                },
                                                "belongProjectCode": {
                                                    "description": "项目归属code",
                                                    "type": "string"
                                                },
                                                "belongProjectId": {
                                                    "description": "项目归属id",
                                                    "type": "integer"
                                                },
                                                "belongProjectName": {
                                                    "description": "项目归属名称",
                                                    "type": "string"
                                                },
                                                "gmtConfirm": {
                                                    "description": "订单确认时间(支付时间)",
                                                    "type": "string"
                                                },
                                                "gmtCreate": {
                                                    "description": "创建时间",
                                                    "type": "string"
                                                },
                                                "gmtExpire": {
                                                    "description": "订单超时时间",
                                                    "type": "string"
                                                },
                                                "gmtModify": {
                                                    "description": "更新时间\t",
                                                    "type": "string"
                                                },
                                                "gmtOrder": {
                                                    "description": "下单时间",
                                                    "type": "string"
                                                },
                                                "linkMan": {
                                                    "description": "订单联系人",
                                                    "type": "string"
                                                },
                                                "linkPhone": {
                                                    "description": "订单联系电话",
                                                    "type": "string"
                                                },
                                                "note": {
                                                    "description": "订单备注",
                                                    "type": "string"
                                                },
                                                "orderSourceName": {
                                                    "description": "下单方式",
                                                    "type": "string"
                                                },
                                                "orderStatus": {
                                                    "description": "订单状态",
                                                    "type": "integer"
                                                },
                                                "payTime": {
                                                    "description": "支付时间",
                                                    "type": "string"
                                                },
                                                "payTypeName": {
                                                    "description": "支付类型",
                                                    "type": "string"
                                                },
                                                "projectOrderNo": {
                                                    "description": "项目主订单流水号",
                                                    "type": "string"
                                                },
                                                "remark": {
                                                    "description": "客服备注",
                                                    "type": "string"
                                                },
                                                "sourceProjectCode": {
                                                    "description": "项目来源code",
                                                    "type": "string"
                                                },
                                                "sourceProjectId": {
                                                    "description": "项目来源id",
                                                    "type": "integer"
                                                },
                                                "sourceProjectName": {
                                                    "description": "项目来源名称",
                                                    "type": "string"
                                                },
                                                "subOrderType": {
                                                    "description": "子订单类型",
                                                    "type": "string"
                                                },
                                                "totalPayPrice": {
                                                    "description": "实际支付金额（单位分）",
                                                    "type": "integer"
                                                },
                                                "totalSalePrice": {
                                                    "description": "订单总销售价（单位分）",
                                                    "type": "integer"
                                                },
                                                "totalSettlementPrice": {
                                                    "description": "订单总成本价（单位分）",
                                                    "type": "integer"
                                                },
                                                "userId": {
                                                    "description": "用户唯一标识",
                                                    "type": "integer"
                                                },
                                                "userIdTypeName": {
                                                    "description": "用户唯一标识类型,0:无用户标示,1:用户表自增id,2:身份证,3:手机号",
                                                    "type": "string"
                                                },
                                                "userLevelName": {
                                                    "description": "用户等级 0:无，1:VIP",
                                                    "type": "string"
                                                },
                                                "userType": {
                                                    "description": "用户类型",
                                                    "type": "string"
                                                },
                                                "subOrderDetailList": {
                                                    "description": "子订单信息",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "cancelReason": {
                                                                "description": "申请取消原因",
                                                                "type": "string"
                                                            },
                                                            "companyCode": {
                                                                "description": "公司编码",
                                                                "type": "integer"
                                                            },
                                                            "companyName": {
                                                                "description": "公司名称",
                                                                "type": "string"
                                                            },
                                                            "industryCode": {
                                                                "description": "行业code",
                                                                "type": "integer"
                                                            },
                                                            "industryName": {
                                                                "description": "行业名称",
                                                                "type": "string"
                                                            },
                                                            "payPrice": {
                                                                "description": "实付价 单位：分",
                                                                "type": "integer"
                                                            },
                                                            "productCode": {
                                                                "description": "商品编码",
                                                                "type": "integer"
                                                            },
                                                            "projectProductName": {
                                                                "description": "项目商品名称",
                                                                "type": "string"
                                                            },
                                                            "projectStoreName": {
                                                                "description": "项目商户名称",
                                                                "type": "string"
                                                            },
                                                            "quality": {
                                                                "description": "数量",
                                                                "type": "integer"
                                                            },
                                                            "salePrice": {
                                                                "description": "销售价 单位：分",
                                                                "type": "integer"
                                                            },
                                                            "settlementMethodCode": {
                                                                "description": "结算方式code(111)",
                                                                "type": "integer"
                                                            },
                                                            "settlementMethodName": {
                                                                "description": "结算方式名称(111)",
                                                                "type": "string"
                                                            },
                                                            "settlementPrice": {
                                                                "description": "成本价 单位：分",
                                                                "type": "integer"
                                                            },
                                                            "storeCode": {
                                                                "description": "门店编码",
                                                                "type": "integer"
                                                            },
                                                            "subOrderNo": {
                                                                "description": "子订单编号",
                                                                "type": "integer"
                                                            },
                                                            "subOrderStatus": {
                                                                "description": "子订单状态",
                                                                "type": "string"
                                                            },
                                                            "subPayStatus": {
                                                                "description": "支付状态(131)",
                                                                "type": "integer"
                                                            },
                                                            "couponResp": {
                                                                "properties": {
                                                                    "couponOrderRespList": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "$ref": "#/components/schemas/couponOrderResp"
                                                                        }
                                                                    },
                                                                    "productCategory": {
                                                                        "description": "产品类别：210001-现金券，210002-商品券，210003-兑换券",
                                                                        "type": "integer"
                                                                    },
                                                                    "productDemandCouponType": {
                                                                        "description": "产品需求券类型：220001-实际券，220002-虚拟券",
                                                                        "type": "integer"
                                                                    },
                                                                    "productLssuanceMethod": {
                                                                        "description": "产品发放方式：211001-券码号，211002-卡号卡密，211003-券链接，211004-直充",
                                                                        "type": "integer"
                                                                    },
                                                                    "providerName": {
                                                                        "description": "渠道名称",
                                                                        "type": "string"
                                                                    },
                                                                    "stockCode": {
                                                                        "description": "库存CODE",
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "discountRespList": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/components/schemas/discountResp"
                                                                }
                                                            },
                                                            "foodResp": {
                                                                "$ref": "#/components/schemas/foodOrderResp"
                                                            },
                                                            "hotelResp": {
                                                                "$ref": "#/components/schemas/hotelOrderResp"
                                                            },
                                                            "spaResp": {
                                                                "$ref": "#/components/schemas/spaResp"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "接口返回的当前unix时间戳",
                                            "type": "integer"
                                        },
                                        "errorInfo": {
                                            "description": "兼容java返回数据，可废弃",
                                            "type": "string"
                                        },
                                        "errorMessage": {
                                            "description": "兼容java返回数据，可废弃",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "couponOrderResp": {
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "用户ID",
                                "type": "integer"
                            },
                            "email": {
                                "description": "email",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "hotelOrderResp": {
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "用户ID",
                                "type": "integer"
                            },
                            "email": {
                                "description": "email",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "foodOrderResp": {
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "用户ID",
                                "type": "integer"
                            },
                            "email": {
                                "description": "email",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "discountResp": {
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "用户ID",
                                "type": "integer"
                            },
                            "email": {
                                "description": "email",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "spaResp": {
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "用户ID",
                                "type": "integer"
                            },
                            "email": {
                                "description": "email",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        }
    }
}